Added TrackGenerator: This class generates a track with no loose ends
Currently the track consists solely of straights. What needs to be done, is to make nice curves at the control points

+++ TrackGenerator/OpenDRIVE_API.py
@@ -65,7 +65,7 @@ def generate(out_path: str) -> None:
     street_length = 0
     for i in reference_line_parts:
         planview.append(i)
-        street_length += int(i.get(key='length'))
+        street_length += float(i.get(key='length'))
     road.set('length', str(street_length))
 
     # Lanes
